Fixed errors in ANSI-C library
git-svn-id: svn+ssh://svn.cprover.org/srv/svn/cbmc/trunk@1608 6afb6bc1-c8e4-404c-8f48-9ae832c5b171

+++ src/ansi-c/library/math.c
@@ -72,11 +72,11 @@ inline int signbit(double d) { return __CPROVER_sign(d); }
 
 /* FUNCTION: __signbitf */
 
-inline float __signbitf(float f) { return __CPROVER_sign(f); }
+inline int __signbitf(float f) { return __CPROVER_sign(f); }
 
 /* FUNCTION: __signbit */
 
-inline double __signbit(double d) { return __CPROVER_sign(d); }
+inline int __signbit(double d) { return __CPROVER_sign(d); }
 
 /* FUNCTION: __fpclassifyd */
